Backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
opencv-python-headless
torch
torchvision
einops
timm
lama-cleaner
pennylane
